requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
azure-ai-inference==1.0.0b6
azure-core==1.30.0
